test(item): migrate a11y
